불필요한 코드 제거 및 폴더 구조 정리
