+++ .gitignore
@@ -1,3 +1,25 @@
+# 환경변수
 .env
+
+# Python 캐시
 __pycache__/
 *.pyc
+*.pyo
+*.pyd
+
+# 가상환경
+venv/
+.venv/
+
+# 이미지 생성 아티팩트
+*.png
+*.jpg
+*.jpeg
+
+# IDE
+.DS_Store
+.vscode/
+.idea/
+
+# 배포
+*.log
