+++ database.py
@@ -4,9 +4,17 @@ from sqlalchemy.orm import sessionmaker
 import os
 from dotenv import load_dotenv
 
-load_dotenv()  # 현재 디렉토리의 .env 파일 읽기
+load_dotenv()
 
 DATABASE_URL = os.getenv('DATABASE_URL', os.environ.get('DATABASE_URL'))
 engine = create_engine(DATABASE_URL, pool_recycle=500)
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 Base = declarative_base()
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
