+++ schemas/psy.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class PsyInput(BaseModel):
+    user_id: int
+    diary_id: int
+    diary_text: str
