+++ schemas/share_diary.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+from datetime import datetime
+
+
+class ShareDiaryData(BaseModel):
+    diary_id: int
+    created_at: datetime
