--- test.py
@@ -1,30 +1,0 @@
-# 제가 생각하는 좋은 방향
-
-# 이미지 생성을 할 때 거기서
-# 국가, 나이, 성별
-
-# 데이터베이스를 건드리고 싶지가 않기때문입니다.
-# 절대 귀찮지는 않거든요 복잡한게 아니라
-
-# 뭔가 꼬이면 지옥이 펼쳐져서 약간의 두려움
-
-
-
-
-# 이미지 생성하기
-# 국가 = 대한민국, 나이 = 26, 성별 = 남
-
-
-# post에 body로 저게 오겠죠
-
-# 그 다음부터
-
-
-
-
-# ---주찬님---
-
-# 테이블을 만들어보자
-
-# 테이블을 하나 만들자
-# 국가, 나이, 성별 -> 어느 시점에 테이블에 저장이 되는건가요
